requirements.txt
chromadb
langchain-core==0.2.4
langchain-text-splitters==0.2.1
openai==1.31.0
PyPDF2==3.0.1
PyYAML==6.0.1